+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Remove multi-tenant proxy restart hack.
+
 ## [0.4.4] - 2024-01-22
 
 ### Changed